requirements.txt
# Web Application Dependencies
Django==4.2.7
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-filter==23.3
django-widget-tweaks==1.5.0
django-rest-framework==0.1.0
psycopg2-binary>=2.9.5
python-dotenv>=0.21.0
Pillow==10.1.0

# Data Analysis and AI
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.0
matplotlib==3.8.2
seaborn==0.13.0

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
linkedin-api==2.0.0
scrapy==2.11.0

# Desktop Application
PyQt6>=6.4.0
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
tqdm==4.66.1 